+++ doc-src/config/all.py
@@ -87,7 +87,7 @@ release = meta.app_version
 
 # List of patterns, relative to source directory, that match files and
 # directories to ignore when looking for source files.
-exclude_patterns = ['_build', 'Thumbs.db', '.DS_Store']
+exclude_patterns = ['_build', 'Thumbs.db', '.DS_Store', 'setup.py']
 
 # The reST default role (used for this markup: `text`) to use for all
 # documents.