+++ test/test_cdau.py
@@ -3,10 +3,14 @@ from __future__ import unicode_literals
 from builtins import str
 import unittest
 import mock
+import os
 
 from catatom2osm import config, cdau, layer
 from catatom2osm.app import QgsSingleton
+from catatom2osm.compat import install_gettext
 qgs = QgsSingleton()
+os.environ['LANGUAGE'] = 'C'
+install_gettext('catato2osm', '')
 
 
 def get_func(f):